Create Jenkinsfile
This is a Jenkins declarative pipeline file (`Jenkinsfile`) that defines a CI/CD (Continuous Integration/Continuous Delivery) workflow using Jenkins and Docker. In this case, the workflow is quite simple and focuses on building a Maven project using a specific Docker image.

Here's the meaning of the code:

1. **Requires the Docker Pipeline plugin:**
   - This comment indicates that this Jenkins file requires the use of the Docker Pipeline plugin. This plugin allows Jenkins to run build steps inside Docker containers.

2. **pipeline { ... }:**
   - This defines a Jenkins pipeline declaration block.

3. **agent { docker { image 'maven:3.9.4-eclipse-temurin-17-alpine' } }:**
   - This `agent` block specifies the environment in which the pipeline stages will run. In this case, Docker is used as the agent, and it defines the Docker image that will be used to run the pipeline stages. The Docker image is named 'maven:3.9.4-eclipse-temurin-17-alpine,' which is an image containing Maven and Java 17 on an Alpine Linux distribution.

4. **stages { ... }:**
   - This block defines the pipeline stages. In this case, there is only one stage called 'build.'

5. **stage('build') { ... }:**
   - This block defines the 'build' stage.

6. **steps { ... }:**
   - This block defines the steps that will be executed in the build stage.

7. **sh 'mvn --version':**
   - This is a shell step (`sh`) that will execute the 'mvn --version' command. This will print the version of Maven to the pipeline output. In summary, this step verifies that Maven is available in the Docker image used and displays its version.

In summary, this Jenkinsfile defines a workflow that uses a specific Docker image with Maven and Java 17 to execute the 'mvn --version' command in a stage named 'build.' This is a very simple example of a pipeline, and in a real-world workflow, you would add more stages and steps to automate tasks like building, testing, and deploying your application.